+++ CHANGELOG.md
@@ -1,5 +1,31 @@
 # Changelog
 
+## v11.1.1 (2026-01-06)
+
+### Bug fixes
+
+- Base notification data on device capabilities ([`1824b72`](https://github.com/chemelli74/aioamazondevices/commit/1824b721a2d5f74786544ead7942748585e56086))
+
+
+## v11.1.0 (2026-01-05)
+
+### Features
+
+- Add support for aqm devices ([`3e47826`](https://github.com/chemelli74/aioamazondevices/commit/3e478267b6b55d5cddb12cbeb584f3e33b123f00))
+
+
+## v11.0.3 (2026-01-05)
+
+### Bug fixes
+
+- Add connectivity to graphql sensors ([`eda51b8`](https://github.com/chemelli74/aioamazondevices/commit/eda51b8fb91a22dbeafcef7625fea1fccee977b5))
+
+
+### Build system
+
+- Bump aiohttp from 3.13.2 to 3.13.3 ([`a064bd3`](https://github.com/chemelli74/aioamazondevices/commit/a064bd359953988e291998d55958b0afad715c90))
+
+
 ## v11.0.2 (2025-12-30)
 
 ### Bug fixes